Move `extension` to `extension_host`
