+++ crates/extension_host/src/extension_builder.rs
