+++ crates/extension_host/src/extension_host.rs
