+++ crates/extension_host/src/extension_lsp_adapter.rs
