+++ crates/extension_host/src/wasm_host.rs
