+++ crates/extension_host/src/wasm_host/wit.rs
