+++ crates/extensions_ui/src/extension_indexed_docs_provider.rs
@@ -7,7 +7,7 @@ use futures::FutureExt;
 use indexed_docs::{IndexedDocsDatabase, IndexedDocsProvider, PackageName, ProviderId};
 use wasmtime_wasi::WasiView;
 
-use extension::wasm_host::{WasmExtension, WasmHost};
+use extension_host::wasm_host::{WasmExtension, WasmHost};
 
 pub struct ExtensionIndexedDocsProvider {
     pub(crate) extension: WasmExtension,
