+++ crates/zed/Cargo.toml
@@ -40,7 +40,7 @@ db.workspace = true
 diagnostics.workspace = true
 editor.workspace = true
 env_logger.workspace = true
-extension.workspace = true
+extension_host.workspace = true
 extensions_ui.workspace = true
 feature_flags.workspace = true
 feedback.workspace = true
